chore: use onAllReady for everyone by default
allows defer + concurrent features to work, but will still buffer the entire response
